requirements.txt
fastapi
uvicorn
requests
websockets
SpeechRecognition
pyttsx3
gTTS
pyautogui
psutil
pywhatkit
selenium
webdriver-manager
python-dotenv
sqlalchemy
pydantic
pyperclip
schedule
pytest
